user_ccounts/requirements.txt
asgiref==3.8.1
certifi==2025.1.31
charset-normalizer==3.4.1
Django==5.2
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
dotenv==0.9.9
drf-yasg==1.21.10
idna==3.10
inflection==0.5.1
mysqlclient==2.2.7
packaging==24.2
psycopg2==2.9.10
PyJWT==2.9.0
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
requests==2.32.3
sqlparse==0.5.3
structlog==25.2.0
structured_log==1.9
uritemplate==4.1.1
urllib3==2.3.0
